+++ test/provider_vars/main.tf
@@ -1,11 +1,12 @@
 variable "conjur_api_key" {}
+variable "conjur_ssl_cert" {}
 
 provider "conjur" {
-  appliance_url = "http://conjur-server"
+  appliance_url = "https://conjur-server"
   account = "myaccount"
   login = "admin"
   api_key = var.conjur_api_key
-  # ssl_cert = "-----BEGIN CERTIFICATE-----..."
+  ssl_cert = var.conjur_ssl_cert
   # ssl_cert_path = "/etc/conjur.pem"
 }